pyproject.toml
[project]
name = "music-scheduling"
version = "0.1.0"
description = ""
authors = [
    {name = "Ben Piticaru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "google-api-python-client (>=2.168.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
